refactor(auth): relocate BlockListSigner to nostr_client (#4741)
auth_service depended on the whole content_blocklist_repository package solely
for the pure BlockListSigner interface (ContentBlocklistRepository itself was
referenced 0 times) — a repo->repo layering smell that would follow auth into
its extracted package. Moves the interface to nostr_client (a lower-level client
abstraction both consumers already depend on). content_blocklist_repository
re-exports it from nostr_client for backward compat; auth_service now imports it
from nostr_client and no longer imports content_blocklist_repository at all.

PR3 (layering prerequisites) for the auth_service extraction (#4338 W2).
