+++ mobile/packages/nostr_client/lib/nostr_client.dart
@@ -5,6 +5,7 @@
 /// and connection handling. Integrates SDK, gateway, and caching.
 library;
 
+export 'src/block_list_signer.dart';
 export 'src/models/models.dart';
 export 'src/nip89_client_tag.dart';
 export 'src/nostr_client.dart';
